requirements.txt
langchain==0.3.27
python-dotenv==1.1.1
ipykernel==6.30.1
langchain-groq==0.3.7
langchain-google-genai==2.1.9
langchain-community==0.3.27
pypdf==5.9.0
faiss-cpu==1.11.0.post1
structlog==25.4.0
PyMuPDF==1.26.3
pandas==2.3.1
streamlit==1.48.1
langchain-core==0.3.72
docx2txt==0.9
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
cfn-lint==1.39.1

-e .